fixed EVERYTHING.
- changed the fact that you made pause menus voids instead of IEnumerators
- fixed pause menu
- fresh pause menu
- fixed animations on pause menu

- made text narrative in single player more readable
- made enemy animation cooler.
- fixed camera fov being 1e-05 because of a translation error between a float and an int in the options menu script.

- fixed sensitivity being set to max in the menu if it's your first time playing the game.

- the first level of singleplayer is 45% done.

+++ Source Code/Assets/Scripts/MenuScripts/MyOptionsMenu.cs
@@ -30,12 +30,12 @@ public class MyOptionsMenu : MonoBehaviour
         //mouse settings
         if (PlayerPrefs.HasKey("Mouse"))
         {
-            MouseLook.value = PlayerPrefs.GetFloat("Mouse");
+            MouseLook.value = PlayerPrefs.GetInt("Mouse");
         }
         else
         {
-            MouseLook.value = 1f;
-            PlayerPrefs.SetFloat("Mouse", MouseLook.value);
+            MouseLook.value = 1;
+            PlayerPrefs.SetInt("Mouse", (int) MouseLook.value);
             PlayerPrefs.Save();
         }
 
@@ -189,7 +189,7 @@ public class MyOptionsMenu : MonoBehaviour
         else
         {
             cameraFOVSlider.value = 60;
-            PlayerPrefs.SetFloat("FOV", cameraFOVSlider.value);
+            PlayerPrefs.SetInt("FOV", (int) cameraFOVSlider.value);
             PlayerPrefs.Save();
         }
